% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aa2codon}
\alias{aa2codon}
\title{amino acids to codons}
\format{
## `aa2codon`
data.frame with two columns: amino_acid, and codons.
}
\source{
It is actually the standard genetic code.
}
\usage{
aa2codon
}
\description{
A data.frame of mapping from amino acids to codons
}
\keyword{datasets}
