% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm6_cds}
\alias{dm6_cds}
\title{fly CDS sequences}
\format{
## `dm6_cds`
A DNAStringSet of 20,476 fly CDS sequences
}
\source{
<https://ftp.ensembl.org/pub/release-107/fasta/drosophila_melanogaster/cds/Drosophila_melanogaster.BDGP6.32.cds.all.fa.gz>
}
\usage{
dm6_cds
}
\description{
CDSs of all protein-coding genes in Drosophila melanogaster
}
\keyword{datasets}
