% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_weight.R
\name{extract_trna_gcn}
\alias{extract_trna_gcn}
\title{get tRNA gene copy number from GtRNADB}
\usage{
extract_trna_gcn(trna_seq)
}
\arguments{
\item{trna_seq}{a fasta file of tRNA sequences from GtRNADB}
}
\value{
a table of tRNA gene copy number for each anticodon. The name of the element 
corresponds to the anticodon and the amino acid it carries. tRNA transporting either 
initiator methionine (iMet) in eukaryotes and archaea, or formylmethionine (fMet) in prokaryotes, 
utilized to initiate translation, is excluded from the count.
}
\description{
\code{extract_trna_gcn} get tRNA gene copy number from GtRNADB
}
\examples{
# get tRNA gene copy number for yeast
trna_gcn <- extract_trna_gcn(yeast_trna)

}
