% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_index.R
\name{get_cai}
\alias{get_cai}
\title{Calculate CAI}
\usage{
get_cai(cf, rscu)
}
\arguments{
\item{cf}{matrix of codon frequencies as calculated by `count_codons()`.}

\item{rscu}{rscu table containing CAI weight for each codon. This table could be
generated with `est_rscu` or prepared manually.}
}
\value{
a named vector of CAI values
}
\description{
\code{get_cai} calculates Codon Adaptation Index (CAI) of each input CDS
}
