% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_index.R
\name{get_tai}
\alias{get_tai}
\title{Calculate TAI}
\usage{
get_tai(cf, trna_w)
}
\arguments{
\item{cf}{matrix of codon frequencies as calculated by `count_codons()`.}

\item{trna_w}{tRNA weight for each codon, can be generated with `est_trna_weight()`.}
}
\value{
a named vector of TAI values
}
\description{
\code{get_tai} calculates tRNA Adaptation Index (TAI) of each CDS
}
