% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_mt}
\alias{human_mt}
\title{human mitochondrial CDS sequences}
\format{
a DNAStringSet of 13 sequences
}
\source{
<https://www.ensembl.org/index.html>
}
\usage{
human_mt
}
\description{
CDSs of 13 protein-coding genes in the human mitochondrial genome extracted from ENSEMBL Biomart
}
\examples{
head(human_mt)
}
\keyword{datasets}
