% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{rev_comp}
\alias{rev_comp}
\title{Reverse complement}
\usage{
rev_comp(seqs)
}
\arguments{
\item{seqs}{input sequences, DNAStringSet or named vector of sequences}
}
\value{
reverse complemented input sequences as a DNAStringSet.
}
\description{
\code{rev_comp} creates reverse complemented version of the input sequence
}
