% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{seq_to_codons}
\alias{seq_to_codons}
\title{Convert CDS to codons}
\usage{
seq_to_codons(seq)
}
\arguments{
\item{seq}{DNAString, or an object that can be coerced to a DNAString}
}
\value{
a character vector of codons
}
\description{
\code{seq_to_codons} converts a coding sequence to a vector of codons
}
