% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{slide}
\alias{slide}
\title{slide window interval generator}
\usage{
slide(from, to, step = 1, before = 0, after = 0)
}
\arguments{
\item{from}{integer, the start of the sequence}

\item{to}{integer, the end of the sequence}

\item{step}{integer, the step size}

\item{before}{integer, the number of values before the center of a window}

\item{after}{integer, the number of values after the center of a window}
}
\value{
data.table with start, center, and end columns
}
\description{
\code{slide} generates a data.table with start, center, and end columns
for a sliding window analysis.
}
\examples{
slide(1, 10, step = 2, before = 1, after = 1)

}
