% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{slide_plot}
\alias{slide_plot}
\title{plot sliding window codon usage}
\usage{
slide_plot(windt, index_name = "Index")
}
\arguments{
\item{windt}{data.table, the sliding window codon usage
generated by `slide_apply`.}

\item{index_name}{character, the name of the index to display.}
}
\value{
ggplot2 plot.
}
\description{
\code{slide_plot} visualizes codon usage in sliding window.
}
\examples{
sw <- slide_apply(yeast_cds[[1]], get_enc, step = 1, before = 10, after = 10)
slide_plot(sw)
}
