% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_weight.R
\name{trna_gcn}
\alias{trna_gcn}
\title{get tRNA gene copy number from GtRNADB}
\usage{
trna_gcn(trna_seq)
}
\arguments{
\item{trna_seq}{a fasta file of tRNA sequences from GtRNADB}
}
\value{
a table of tRNA gene copy number for each anticodon
}
\description{
\code{trna_gcn} get tRNA gene copy number from GtRNADB
}
\examples{
# get tRNA gene copy number for yeast
trna_gcn <- trna_gcn(yeast_trna)

}
