% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_trna}
\alias{yeast_trna}
\title{yeast tRNA sequences}
\format{
a RNAStringSet with a length of 275.
}
\source{
<http://gtrnadb.ucsc.edu/genomes/eukaryota/Scere3/sacCer3-mature-tRNAs.fa>
}
\usage{
yeast_trna
}
\description{
Yeast tRNA sequences obtained from gtRNAdb.
}
\examples{
yeast_trna
}
\references{
Chan PP, Lowe TM. 2016. GtRNAdb 2.0: an expanded database of transfer RNA genes identified in complete and draft genomes. Nucleic Acids Res 44:D184-189.
}
\keyword{datasets}
