% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_trna_gcn}
\alias{yeast_trna_gcn}
\title{yeast tRNA gene copy numbers (GCN)}
\format{
## `yeast_trna_gcn`
A named (anticodon) vector with a length of 41 anticodons.
}
\source{
<http://gtrnadb.ucsc.edu/genomes/eukaryota/Scere3/sacCer3-mature-tRNAs.fa>
}
\usage{
yeast_trna_gcn
}
\description{
Yeast tRNA gene copy numbers (GCN) by anticodon obtained from gtRNAdb.
}
\references{
Chan, P.P. & Lowe, T.M. (2016) GtRNAdb 2.0: an expanded database of transfer RNA genes identified in complete and draft genomes. Nucl. Acids Res. 44(Database issue):D184-D189.Chan, P.P. & Lowe, T.M. (2016) GtRNAdb 2.0: an expanded database of transfer RNA genes identified in complete and draft genomes. Nucl. Acids Res. 44(Database issue):D184-D189.
}
\keyword{datasets}
