% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{count_codons}
\alias{count_codons}
\title{Count occurrences of different codons}
\usage{
count_codons(seqs, ...)
}
\arguments{
\item{seqs}{CDS sequences, DNAStringSet}
}
\value{
matrix of codon (column) frequencies of each CDS (row)
}
\description{
\code{count_codons} tabulates the occurrences of all the 64 codons in input CDSs
}
