% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_index.R
\name{get_gc}
\alias{get_gc}
\title{GC contents}
\usage{
get_gc(cf)
}
\arguments{
\item{cf}{matrix of codon frequencies as calculated by `count_codons()`.}
}
\value{
a named vector of GC contents.
}
\description{
Calculate GC content of the whole sequences.
}
\examples{
# estimate GC content of yeast genes
cf_all <- count_codons(yeast_cds)
gc <- get_gc(cf_all)
head(gc)
hist(gc)

}
