% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_index.R
\name{get_tai}
\alias{get_tai}
\title{Calculate TAI}
\usage{
get_tai(cf, trna_w, w_format = "cubar")
}
\arguments{
\item{cf}{matrix of codon frequencies as calculated by \code{count_codons()}. 
The start codons of CDSs need to be removed.}

\item{trna_w}{tRNA weight for each codon, can be generated with \code{est_trna_weight()}.}

\item{w_format}{"cubar" or "tAI". tRNA weight values from cubar (default) or the package tAI.}
}
\value{
a named vector of TAI values. The names of the elements correspond to the sequence names.
}
\description{
\code{get_tai} calculates tRNA Adaptation Index (TAI) of each CDS
}
\examples{
# calculate TAI of yeast genes based on genomic tRNA copy numbers
w <- est_trna_weight(yeast_trna_gcn)
yeast_cds_qc <- check_cds(yeast_cds)
cf <- count_codons(yeast_cds_qc)
tai <- get_tai(cf, w)
head(tai)
hist(tai)

}
\references{
dos Reis M, Savva R, Wernisch L. 2004. Solving the riddle of codon usage
  preferences: a test for translational selection. Nucleic Acids Res 32:5036-5044.
}
