% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_info.R
\name{show_codon_tables}
\alias{show_codon_tables}
\title{show available codon tables}
\usage{
show_codon_tables()
}
\value{
NULL.
}
\description{
\code{show_codon_tables} print a table of available genetic code from NCBI through
`Biostrings::GENETIC_CODE_TABLE`.
}
