% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_cds}
\alias{yeast_cds}
\title{yeast CDS sequences}
\format{
a DNAStringSet of 6600 sequences
}
\source{
<https://ftp.ensembl.org/pub/release-107/fasta/saccharomyces_cerevisiae/cds/Saccharomyces_cerevisiae.R64-1-1.cds.all.fa.gz>
}
\usage{
yeast_cds
}
\description{
CDSs of all protein-coding genes in Saccharomyces_cerevisiae
}
\examples{
head(yeast_cds)
}
\keyword{datasets}
