% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_exp}
\alias{yeast_exp}
\title{yeast mRNA expression levels}
\format{
a data.frame with 6717 rows and three columns:
\describe{
\item{gene_id}{gene ID}
\item{gene_name}{gene name}
\item{fpkm}{mRNA expression level in Fragments per kilobase per million reads}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57385>
}
\usage{
yeast_exp
}
\description{
Yeast mRNA FPKM determined from rRNA-depleted (RiboZero) total RNA-Seq libraries. RUN1_0_WT and RUN2_0_WT (0 min after RNA Pol II repression) were averaged and used here.
}
\examples{
head(yeast_exp)
}
\references{
Presnyak V, Alhusaini N, Chen YH, Martin S, Morris N, Kline N, Olson S, Weinberg D, Baker KE, Graveley BR, et al. 2015. Codon optimality is a major determinant of mRNA stability. Cell 160:1111-1124.
}
\keyword{datasets}
