% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_half_life}
\alias{yeast_half_life}
\title{Half life of yeast mRNAs}
\format{
a data.frame with 3888 rows and three columns:
\describe{
\item{gene_id}{gene id}
\item{gene_name}{gene name}
\item{half_life}{mRNA half life in minutes}
}
}
\source{
<https://doi.org/10.1016/j.cell.2015.02.029>
}
\usage{
yeast_half_life
}
\description{
Half life of yeast mRNAs in Saccharomyces_cerevisiae calculated from rRNA-deleted total RNAs by Presnyak et al.
}
\examples{
head(yeast_half_life)
}
\references{
Presnyak V, Alhusaini N, Chen YH, Martin S, Morris N, Kline N, Olson S, Weinberg D, Baker KE, Graveley BR, et al. 2015. Codon optimality is a major determinant of mRNA stability. Cell 160:1111-1124.
}
\keyword{datasets}
