% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_half_life}
\alias{yeast_half_life}
\title{Half life of yeast mRNAs}
\format{
## `yeast_half_life`
A data.frame with 3888 rows and three columns (gene_id, gene_name, half_life in min)
}
\source{
<https://doi.org/10.1016/j.cell.2015.02.029>
}
\usage{
yeast_half_life
}
\description{
Half life of yeast mRNAs in Saccharomyces_cerevisiae calculated from rRNA-deleted total RNAs by Presnyak et al.
}
\keyword{datasets}
